requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.8.0; python_version >= '3.9'
atproto==0.0.58; python_version < '3.14' and python_version >= '3.8'
boto3==1.37.17; python_version >= '3.8'
botocore==1.37.17; python_version >= '3.8'
cbor2==5.6.5; python_version >= '3.8'
certifi==2025.1.31; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
click==8.1.8; python_version >= '3.7'
contourpy==1.3.1; python_version >= '3.10'
cryptography==43.0.3; python_version >= '3.7'
cycler==0.12.1; python_version >= '3.8'
dnspython==2.7.0; python_version >= '3.9'
duckdb==1.2.0; python_full_version >= '3.7.0'
et-xmlfile==2.0.0; python_version >= '3.8'
fonttools==4.56.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.27.2; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
imbalanced-learn==0.13.0; python_version >= '3.10'
imblearn==0.0
jmespath==1.0.1; python_version >= '3.7'
joblib==1.5.0; python_version >= '3.9'
kiwisolver==1.4.8; python_version >= '3.10'
libipld==3.0.0; python_version >= '3.8'
matplotlib==3.10.1; python_version >= '3.10'
numpy==2.2.5; python_version >= '3.10'
openpyxl==3.1.5; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pillow==11.1.0; python_version >= '3.9'
pycparser==2.22; python_version >= '3.8'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pyparsing==3.2.1; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2025.1
s3transfer==0.11.4; python_version >= '3.8'
scikit-learn==1.6.1; python_version >= '3.9'
scipy==1.15.2; python_version >= '3.10'
seaborn==0.13.2; python_version >= '3.8'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sklearn-compat==0.1.3; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
threadpoolctl==3.6.0; python_version >= '3.9'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2025.1; python_version >= '2'
urllib3==2.3.0; python_version >= '3.9'
websockets==13.1; python_version >= '3.8'
xgboost==3.0.0; python_version >= '3.10'
